pyproject.toml
[project]
name = "tunarr-playlists"
version = "0.1.0"
description = "Sync Plex playlists to Tunarr channels"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "plexapi>=4.15.0",
    "beautifulsoup4>=4.12.0",
    "pyyaml>=6.0.0",
    "cloudscraper>=1.2.71",
]

[project.scripts]
tunarr-sync = "tunarr_playlists.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
